requirements37.txt
certifi==2021.10.8
Cython
flatbuffers==2.0
loguru==0.5.3
numpy==1.19.2
onnxruntime==1.9.0
opencv-python==4.3.0.36
Pillow==9.5.0
protobuf==3.19.1
pyclipper==1.3.0.post5
PyQt5==5.15.2
PyQt5-sip==12.9.0
PySide2==5.15.2
shapely==2.0.4
shiboken2==5.15.2
tqdm==4.61.1
